+++ components/select/Select.js
@@ -34,7 +34,8 @@ class Select extends Component {
     optionWidth: PropTypes.number,
     style: PropTypes.object,
     onChange: PropTypes.func,
-    dropdownRender: PropTypes.func
+    dropdownRender: PropTypes.func,
+    open: PropTypes.bool
   }
 
   static defaultProps = {
@@ -47,7 +48,8 @@ class Select extends Component {
     autoload: false,
     placeholder: '请选择',
     noFoundTip: '无内容',
-    showCheckAll: false
+    showCheckAll: false,
+    open: true
   }
 
   constructor (props) {
@@ -100,7 +102,8 @@ class Select extends Component {
   }
 
   clickOutside (e) {
-    if (ReactDOM.findDOMNode(this.selectInput) && ReactDOM.findDOMNode(this.selectInput).contains(e.target)) {
+    const selectInput = ReactDOM.findDOMNode(this.selectInput)
+    if (selectInput && selectInput.contains(e.target)) {
       return
     }
     this.hideDropdown()
@@ -191,35 +194,32 @@ class Select extends Component {
     this.onClickOption(item, focusedIndex)
   }
 
-  onChange () {
-    const {
-      selectedItems
-    } = this.state
+  onChange (changedItems) {
+    const { selectedItems } = this.state
 
-    this.props.onChange && this.props.onChange(selectedItems)
+    this.props.onChange && this.props.onChange(selectedItems, changedItems)
   }
 
-  checkAll (e) { // 全选
+  checkAll (e) {
+    // 全选
     e && e.stopPropagation()
 
-    const {
-      dropdownItems
-    } = this.state
-    let selectedItems = this.state.selectedItems.concat()
-
+    const { dropdownItems, selectedItems } = this.state
+    let _selectedItems = [...selectedItems]
+    let changedItems = []
     dropdownItems.forEach(item => {
       if (!item.disabled && this.matchFilter(item)) {
-        let itemIndex = selectedItems.findIndex((sItem) => {
-          return sItem.id === item.id
-        })
-        itemIndex === -1 && selectedItems.push(item)
+        if (!_selectedItems.map(selectItem => selectItem.id).includes(item.id)) {
+          _selectedItems.push(item)
+          changedItems.push(item)
+        }
       }
     })
     this.setState({
-      selectedItems
+      selectedItems: _selectedItems
     }, () => {
       this.selectInput.focus()
-      this.onChange()
+      this.onChange(changedItems)
     })
   }
 
@@ -253,7 +253,7 @@ class Select extends Component {
         this.clearKeyword() // 多选状态清空筛选
       }
 
-      this.onChange()
+      this.onChange(item)
     })
   }
 
@@ -265,7 +265,7 @@ class Select extends Component {
     })
   }
 
-  handleInputClick (e) {
+  handleInputClick = (e) => {
     let {
       dropdownShow
     } = this.state
@@ -309,18 +309,18 @@ class Select extends Component {
       selectedItems
     }, () => {
       this.selectInput.focus()
-      this.onChange()
+      this.onChange(item)
     })
   }
   // 全部删除
   deleteAllItems () {
     const focusedIndex = this.resetFocusedIndex()
-
+    const changedItems = [...this.state.selectedItems]
     this.setState({
       focusedIndex,
       selectedItems: []
     }, () => {
-      this.onChange()
+      this.onChange(this.props.mode === 'multiple' ? changedItems : changedItems[0])
       this.onFilterItems('')
     })
   }
@@ -513,7 +513,11 @@ class Select extends Component {
             multipleMode={multipleMode}
             container={this.selectInputContainer}
             moveFocusedIndex={this.moveFocusedIndex.bind(this)}
-            onClick={this.handleInputClick.bind(this)}
+            onClick={()=>{
+              if(this.props.open) {
+                this.handleInputClick()
+              }
+            }}
             onDelete={this.deleteItem.bind(this)}
             onClear={this.deleteAllItems.bind(this)}
             onSearch={this.debouncedFilterItems.bind(this)}
@@ -522,7 +526,7 @@ class Select extends Component {
         </div>
         { children }
         <Popper
-          show={dropdownShow}
+          show={dropdownShow && this.props.open}
           attachEle={this.selectInputContainer}
           zIndex={1050}
           topGap={5}