+++ site/locales/zh-CN.js
@@ -42,7 +42,12 @@ module.exports = {
     grid: 'Grid 栅格',
     stepper: 'Stepper 步骤',
     icon: 'Icon 图标',
-    progress: 'Progress 进度条'
+    progress: 'Progress 进度条',
+    card: 'Card  卡片',
+    timeline: 'Timeline 时间轴',
+    transfer: 'Transfer 穿梭框',
+    switch: 'Switch 开关',
+    rate: 'Rate 评分'
   },
   designs: {
     'design-patterns': '设计模式',