+++ site/pages/components/index.js
@@ -32,8 +32,8 @@ export default {
       'checkbox': require('./checkbox'),
       'date-picker': require('./date-picker'),
       'time-picker': require('./time-picker'),
-      'upload': require('./upload')
-
+      'upload': require('./upload'),
+      'rate': require('./rate')
     },
     'group-data': {
       'table': require('./table'),
@@ -42,7 +42,11 @@ export default {
       'collapse': require('./collapse'),
       'tooltip': require('./tooltip'),
       'popover': require('./popover'),
-      'progress': require('./progress')
+      'progress': require('./progress'),
+      'card': require('./card'),
+      'timeline': require('./timeline'),
+      'transfer': require('./transfer'),
+      'switch': require('./switch')
     },
     'group-tips': {
       'modal': require('./modal'),